+++ aws/route53/outputs.tf
@@ -1,0 +1,3 @@
+output "zones" {
+  value = aws_route53_zone.zone
+}