+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.4.0 2020-07-06
+
+Added AWS Route53 module, can be used to create Route53 zones and records. Supports Route53 alias records as well. 
+
 ## v0.3.0 2020-07-06
 
-Added AWS ACM module, can be used to create and validate SSL certificates. It does the validation through DNS record. 
+Added AWS ACM module, can be used to create and validate SSL certificates. It does the validation through DNS record.
+